Move kinematic solver into internal folder
