+++ Assets/Code/Common/Physics/KinematicBody2D.cs
@@ -3,6 +3,7 @@ using System.Diagnostics.Contracts;
 using UnityEngine;
 using UnityEditor;
 using PQ.Common.Extensions;
+using PQ.Common.Physics.Internal;
 
 
 namespace PQ.Common.Physics
